Addressing code review comments `Dog`, `Request` routes
